requirements.txt
EbookLib>=0.18
python-dotenv>=1.0.0
chardet>=5.0.0 